component styled * renamed files to pascal case
