+++ packages/app/src/components/familyInformation/styles.module.css
@@ -1,0 +1,36 @@
+.main {
+  border-radius: 10px;
+  background-color: white;
+  text-align: left;
+}
+
+/**********/
+/* header */
+/**********/
+
+.header {
+  font-size: 1.2em;
+  border-bottom: 1px solid gray;
+  padding: 1rem 2rem;
+}
+
+/***********/
+/* content */
+/***********/
+
+.content > div > span:last-child,
+.main > div > span:last-child {
+  font-weight: bold;
+}
+
+.content > div {
+  margin: 1rem 2rem;
+  padding-bottom: 1rem;
+}
+.content > div:not(:last-child) {
+  border-bottom: 1px solid gray;
+}
+
+span {
+  display: block;
+}
